make gc operations run on cyclic batches.
