+++ pstoreds/addr_book.go
@@ -4,16 +4,13 @@ import (
 	"context"
 	"fmt"
 	"sort"
-	"strconv"
 	"sync"
-	"sync/atomic"
 	"time"
 
 	ds "github.com/ipfs/go-datastore"
 	query "github.com/ipfs/go-datastore/query"
 	logging "github.com/ipfs/go-log"
 
-	pool "github.com/libp2p/go-buffer-pool"
 	peer "github.com/libp2p/go-libp2p-peer"
 	pstore "github.com/libp2p/go-libp2p-peerstore"
 	pb "github.com/libp2p/go-libp2p-peerstore/pb"
@@ -36,10 +33,12 @@ var (
 	// Peer addresses are stored under the following db key pattern:
 	// /peers/addrs/<b32 peer id no padding>
 	addrBookBase = ds.NewKey("/peers/addrs")
-	// GC lookahead entries are stored in keys with pattern:
-	// /peers/gc/addrs/<unix timestamp of next visit>/<peer ID b32> => nil
-	gcLookaheadBase = ds.NewKey("/peers/gc/addrs")
 )
+
+type dsWriter interface {
+	Put(key ds.Key, value []byte) error
+	Delete(key ds.Key) error
+}
 
 // addrsRecord decorates the AddrBookRecord with locks and metadata.
 type addrsRecord struct {
@@ -50,19 +49,17 @@ type addrsRecord struct {
 
 // Flush writes the record to the datastore by calling ds.Put, unless the record is
 // marked for deletion, in which case the deletion is executed via ds.Delete.
-func (r *addrsRecord) Flush(ds ds.Datastore) (err error) {
+func (r *addrsRecord) Flush(ds dsWriter) (err error) {
 	key := addrBookBase.ChildString(b32.RawStdEncoding.EncodeToString([]byte(r.Id.ID)))
 	if len(r.Addrs) == 0 {
 		return ds.Delete(key)
 	}
-	data := pool.Get(r.Size())
-	defer pool.Put(data)
 
-	n, err := r.MarshalTo(data)
+	data, err := r.Marshal()
 	if err != nil {
 		return err
 	}
-	if err = ds.Put(key, data[:n]); err != nil {
+	if err = ds.Put(key, data); err != nil {
 		return err
 	}
 	// write succeeded; record is no longer dirty.
@@ -130,7 +127,7 @@ type dsAddrBook struct {
 	opts Options
 
 	cache       cache
-	ds          ds.Datastore
+	ds          ds.Batching
 	subsManager *pstoremem.AddrSubManager
 
 	flushJobCh chan *addrsRecord
@@ -145,7 +142,7 @@ var _ pstore.AddrBook = (*dsAddrBook)(nil)
 
 // NewAddrBook initializes a new address book given a Datastore instance, a context for managing the TTL manager,
 // and the interval at which the TTL manager should sweep the Datastore.
-func NewAddrBook(ctx context.Context, store ds.TxnDatastore, opts Options) (ab *dsAddrBook, err error) {
+func NewAddrBook(ctx context.Context, store ds.Batching, opts Options) (ab *dsAddrBook, err error) {
 	var cache cache = new(noopCache)
 	if opts.CacheSize > 0 {
 		if cache, err = lru.NewARC(int(opts.CacheSize)); err != nil {
@@ -282,200 +279,6 @@ func (ab *dsAddrBook) background() {
 			}
 		}
 	}()
-}
-
-var purgeQuery = query.Query{Prefix: gcLookaheadBase.String(), KeysOnly: true}
-
-// purgeCycle runs a single GC cycle, operating within the lookahead window.
-//
-// It scans the lookahead region for entries that need to be visited, and performs a Clean() on them. An errors trigger
-// the removal of the GC entry, in order to prevent unactionable items from accumulating. If the error happened to be
-// temporary, the entry will be revisited in the next lookahead window.
-func (ab *dsAddrBook) purgeCycle() {
-	if atomic.LoadInt32(&ab.gcLookaheadRunning) > 0 {
-		// yield if lookahead is running.
-		return
-	}
-
-	var id peer.ID
-	record := &addrsRecord{AddrBookRecord: &pb.AddrBookRecord{}}
-
-	// This function drops an unparseable GC entry; this is for safety. It is an escape hatch in case
-	// we modify the format of keys going forward. If a user runs a new version against an old DB,
-	// if we don't clean up unparseable entries we'll end up accumulating garbage.
-	dropInError := func(key ds.Key, err error, msg string) {
-		if err != nil {
-			log.Warningf("failed while %s with GC key: %v, err: %v", msg, key, err)
-		}
-		if err = ab.ds.Delete(key); err != nil {
-			log.Warningf("failed to delete corrupt GC lookahead entry: %v, err: %v", key, err)
-		}
-	}
-
-	// This function drops a GC key if the entry is cleaned correctly. It may reschedule another visit
-	// if the next earliest expiry falls within the current window again.
-	dropOrReschedule := func(key ds.Key, ar *addrsRecord) {
-		if err := ab.ds.Delete(key); err != nil {
-			log.Warningf("failed to delete lookahead entry: %v, err: %v", key, err)
-		}
-
-		// re-add the record if it needs to be visited again in this window.
-		if len(ar.Addrs) != 0 && ar.Addrs[0].Expiry <= ab.gcCurrWindowEnd {
-			gcKey := gcLookaheadBase.ChildString(fmt.Sprintf("%d/%s", ar.Addrs[0].Expiry, key.Name()))
-			if err := ab.ds.Put(gcKey, []byte{}); err != nil {
-				log.Warningf("failed to add new GC key: %v, err: %v", gcKey, err)
-			}
-		}
-	}
-
-	results, err := ab.ds.Query(purgeQuery)
-	if err != nil {
-		log.Warningf("failed while fetching entries to purge: %v", err)
-		return
-	}
-	defer results.Close()
-
-	now := time.Now().Unix()
-
-	// keys: 	/peers/gc/addrs/<unix timestamp of next visit>/<peer ID b32>
-	// values: 	nil
-	for result := range results.Next() {
-		gcKey := ds.RawKey(result.Key)
-
-		ts, err := strconv.ParseInt(gcKey.Parent().Name(), 10, 64)
-		if err != nil {
-			dropInError(gcKey, err, "parsing timestamp")
-			log.Warningf("failed while parsing timestamp from key: %v, err: %v", result.Key, err)
-			continue
-		} else if ts > now {
-			// this is an ordered cursor; when we hit an entry with a timestamp beyond now, we can break.
-			break
-		}
-
-		idb32, err := b32.RawStdEncoding.DecodeString(gcKey.Name())
-		if err != nil {
-			dropInError(gcKey, err, "parsing peer ID")
-			log.Warningf("failed while parsing b32 peer ID from key: %v, err: %v", result.Key, err)
-			continue
-		}
-
-		id, err = peer.IDFromBytes(idb32)
-		if err != nil {
-			dropInError(gcKey, err, "decoding peer ID")
-			log.Warningf("failed while decoding peer ID from key: %v, err: %v", result.Key, err)
-			continue
-		}
-
-		// if the record is in cache, we clean it and flush it if necessary.
-		if e, ok := ab.cache.Peek(id); ok {
-			cached := e.(*addrsRecord)
-			cached.Lock()
-			if cached.Clean() {
-				if err = cached.Flush(ab.ds); err != nil {
-					log.Warningf("failed to flush entry modified by GC for peer: &v, err: %v", id.Pretty(), err)
-				}
-			}
-			dropOrReschedule(gcKey, cached)
-			cached.Unlock()
-			continue
-		}
-
-		record.Reset()
-
-		// otherwise, fetch it from the store, clean it and flush it.
-		entryKey := addrBookBase.ChildString(gcKey.Name())
-		val, err := ab.ds.Get(entryKey)
-		if err != nil {
-			// captures all errors, including ErrNotFound.
-			dropInError(gcKey, err, "fetching entry")
-			continue
-		}
-		err = record.Unmarshal(val)
-		if err != nil {
-			dropInError(gcKey, err, "unmarshalling entry")
-			continue
-		}
-		if record.Clean() {
-			err = record.Flush(ab.ds)
-			if err != nil {
-				log.Warningf("failed to flush entry modified by GC for peer: &v, err: %v", id.Pretty(), err)
-			}
-		}
-		dropOrReschedule(gcKey, record)
-	}
-}
-
-var populateLookaheadQuery = query.Query{Prefix: addrBookBase.String(), KeysOnly: true}
-
-// populateLookahead populates the lookahead window by scanning the entire store and picking entries whose earliest
-// expiration falls within the new window.
-//
-// Those entries are stored in the lookahead region in the store, indexed by the timestamp when they need to be
-// visited, to facilitate temporal range scans.
-func (ab *dsAddrBook) populateLookahead() {
-	if !atomic.CompareAndSwapInt32(&ab.gcLookaheadRunning, 0, 1) {
-		return
-	}
-
-	until := time.Now().Add(ab.opts.GCLookaheadInterval).Unix()
-
-	var id peer.ID
-	record := &addrsRecord{AddrBookRecord: &pb.AddrBookRecord{}}
-	results, err := ab.ds.Query(populateLookaheadQuery)
-	if err != nil {
-		log.Warningf("failed while filling lookahead GC region: %v", err)
-		return
-	}
-	defer results.Close()
-
-	for result := range results.Next() {
-		idb32 := ds.RawKey(result.Key).Name()
-		k, err := b32.RawStdEncoding.DecodeString(idb32)
-		if err != nil {
-			log.Warningf("failed while decoding peer ID from key: %v, err: %v", result.Key, err)
-			continue
-		}
-		if id, err = peer.IDFromBytes(k); err != nil {
-			log.Warningf("failed while decoding peer ID from key: %v, err: %v", result.Key, err)
-		}
-
-		// if the record is in cache, use the cached version.
-		if e, ok := ab.cache.Peek(id); ok {
-			cached := e.(*addrsRecord)
-			cached.RLock()
-			if len(cached.Addrs) == 0 || cached.Addrs[0].Expiry > until {
-				cached.RUnlock()
-				continue
-			}
-			gcKey := gcLookaheadBase.ChildString(fmt.Sprintf("%d/%s", cached.Addrs[0].Expiry, idb32))
-			if err = ab.ds.Put(gcKey, []byte{}); err != nil {
-				log.Warningf("failed while inserting GC entry for peer: %v, err: %v", id.Pretty(), err)
-			}
-			cached.RUnlock()
-			continue
-		}
-
-		record.Reset()
-
-		val, err := ab.ds.Get(ds.RawKey(result.Key))
-		if err != nil {
-			log.Warningf("failed which getting record from store for peer: %v, err: %v", id.Pretty(), err)
-			continue
-		}
-		if err := record.Unmarshal(val); err != nil {
-			log.Warningf("failed while unmarshalling record from store for peer: %v, err: %v", id.Pretty(), err)
-			continue
-		}
-		if len(record.Addrs) > 0 && record.Addrs[0].Expiry <= until {
-			gcKey := gcLookaheadBase.ChildString(fmt.Sprintf("%d/%s", record.Addrs[0].Expiry, idb32))
-			if err = ab.ds.Put(gcKey, []byte{}); err != nil {
-				log.Warningf("failed while inserting GC entry for peer: %v, err: %v", id.Pretty(), err)
-			}
-		}
-	}
-
-	ab.gcCurrWindowEnd = until
-	atomic.StoreInt32(&ab.gcLookaheadRunning, 0)
 }
 
 // AddAddr will add a new address if it's not already in the AddrBook.
